+++ TinyNvidiaUpdateChecker/Handlers/ConfigurationHandler.cs
@@ -75,6 +75,8 @@ namespace TinyNvidiaUpdateChecker.Handlers
         /// <summary>
         /// Reads setting from configuration file, and adds if requested key / value is missing - returns a string.</summary>
         /// <param name="key"> Config key to read value from.</param>
+        /// <param name="data"> Optional data to use if setting is not found.</param>
+        /// <param name="setupIfNotFound"> Whether to setup the setting if not found.</param>
         public static string ReadSetting(string key, dynamic data = null, bool setupIfNotFound = true)
         {
             string result = null;